Use a deque for _onceWrittenSymbolsIndices
Change the definition of the _onceWrittenSymbolsIndices field of the
TR_UseDefInfo::AuxiliaryData structure to use a TR::deque rather than a
CS2::ArrayOf for its implementation.  Move the specification for its
required size into its initialization as part of the construction of
AuxiliaryDaya, replacing the currently separate resize operation that
occurs post-construction.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>